fix(表结构): 修改表结构
端点配置表添加icon图标（非必填选项-用于在树结构中的展示）

+++ database/tables/t_sys_permission.sql
@@ -33,7 +33,7 @@ CREATE TABLE `t_sys_permission`
     `menu_type`            int                                                           DEFAULT NULL COMMENT '菜单类型(0:一级菜单; 1:子菜单:2:按钮权限)',
     `perms`                varchar(255) CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci DEFAULT NULL COMMENT '菜单权限编码',
     `perms_type`           varchar(10) CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci  DEFAULT 0 COMMENT '权限策略1显示2禁用',
-    `sort_no`              int                                                          NOT NULL COMMENT '菜单排序',
+    `sort_no`              double(8,2)                                                   NOT NULL COMMENT '菜单排序',
     `always_show`          tinyint(1)                                                    DEFAULT NULL COMMENT '聚合子路由: 1是0否',
     `icon`                 varchar(100) CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci DEFAULT NULL COMMENT '菜单图标',
     `is_route`             tinyint(1)                                                    DEFAULT 1 COMMENT '是否路由菜单: 0:不是  1:是（默认值1）',
